backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
google-generativeai
chromadb
websockets
regex
